elevator_system/requirements.txt
django
djangorestframework
django-redis